split reports into modules
